# api_specification.yml
openapi: 3.0.0
info:
  title: Library System API
  version: 1.0.0
  description: API for managing libraries, books, bookings, ratings, and users.
servers:
  - url: http://localhost:8080/api
tags:
  - name: book
  - name: library
  - name: booking
  - name: rating
  - name: user

paths:
  /libraries/{id}/books:
    get:
      tags:
        - library
      summary: Get all books in a library
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: List of books
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ListOfBooks'

  /libraries:
    get:
      tags:
        - library
      summary: Get all libraries
      responses:
        '200':
          description: List of books
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Library'


  /books/{id}/ratings:
    get:
      tags:
        - book
        - rating
      summary: Get ratings for a book
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: List of ratings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Rating'

  /books/{id}/edit:
    get:
      tags:
        - book
      summary: Get the fields of the book a user can edit
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Editable fields
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/BookEdit'


  /books:
    post:
      tags:
        - book
      summary: Create a new book
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Book'
      responses:
        '201':
          description: Book created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookId'


  /books/{id}:
    get:
      tags:
        - book
      summary: Get book details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: book details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookMitId'

    put:
      tags:
        - book
      summary: Update a book
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BookEdit'
      responses:
        '200':
          description: Book updated

    delete:
      tags:
        - book
      summary: Delete a book
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Book deleted

  /bookings:
    post:
      tags:
        - booking
      summary: Create a new booking
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Booking'
      responses:
        '201':
          description: Booking created

  /bookings/{id}:
    put:
      tags:
        - booking
      summary: Update a booking
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BookingEdit'
      responses:
        '200':
          description: Booking updated
    delete:
      tags:
        - booking
      summary: Delete a booking
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Booking deleted

  /bookings/{id}/edit:
    get:
      tags:
        - booking
      summary: Get editable fields of a specific booking
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Editable fields
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookingEditResponse'

  /ratings:
    post:
      tags:
        - rating
      summary: Submit a new rating
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Rating'
      responses:
        '201':
          description: Rating submitted

  /ratings/{id}:
    put:
      tags:
        - rating
      summary: Edit a rating
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RatingEditRequest'
      responses:
        '200':
          description: Rating updated
    delete:
      tags:
        - rating
      summary: Delete a rating
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Rating deleted

  /ratings/{id}/edit:
    get:
      tags:
        - rating
      summary: get editable field of a rating
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: editable fields
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RatingEditResponse'


  /users:
    get:
      tags:
        - user
      summary: Get user details
      responses:
        '200':
          description: User details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'

  /users/students:
    get:
      tags:
        - user
      summary: Get user details of students
      responses:
        '200':
          description: User details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserStudent'

components:
  schemas:
    Book:
      type: object
      properties:
        libraryId:
          type: integer
          format: int64
        author:
          type: string
        title:
          type: string
        isbn:
          type: string
        image:
          type: string
          format: binary
        publisher:
          type: string
        shortDescription:
          type: string
        publishedYear:
          type: integer

    BookMitId:
      type: object
      properties:
        id:
          type: integer
          format: int64
        libraryId:
          type: integer
          format: int64
        author:
          type: string
        title:
          type: string
        isbn:
          type: string
        image:
          type: string
          format: binary
        publisher:
          type: string
        shortDescription:
          type: string
        publishedYear:
          type: integer
        available:
          type: boolean
        bookingId:
          type: integer
          format: int64

    BookEdit:
      type: object
      properties:
        author:
          type: string
        title:
          type: string
        isbn:
          type: string
        image:
          type: string
          format: binary
        publisher:
          type: string
        shortDescription:
          type: string
        publishedYear:
          type: integer

    ListOfBooks:
      type: object
      properties:
        id:
          type: integer
          format: int64
        author:
          type: string
        title:
          type: string
        isbn:
          type: string
        image:
          type: string
          format: binary
        available:
          type: boolean

    Booking:
      type: object
      properties:
        book_id:
          type: integer
          format: int64
        user_id:
          type: integer
          format: int64
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date

    BookingEdit:
      type: object
      properties:
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date

    BookingEditResponse:
      type: object
      properties:
        user_id:
          type: integer
          format: int64
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date

    Rating:
      type: object
      properties:
        bookId:
          type: integer
          format: int64
        userId:
          type: integer
          format: int64
        rating:
          type: integer
          minimum: 1
          maximum: 5
        comment:
          type: string

    RatingEditResponse:
      type: object
      properties:
        userFullName:
          type: string
        rating:
          type: integer
          minimum: 1
          maximum: 5
        comment:
          type: string

    RatingEditRequest:
      type: object
      properties:
        rating:
          type: integer
          minimum: 1
          maximum: 5
        comment:
          type: string

    User:
      type: object
      properties:
        id:
          type: integer
          format: int64
        userFullName:
          type: string
        role:
          type: string
          enum: [user, librarian]

    UserStudent:
      type: object
      properties:
        id:
          type: integer
          format: int64
        userFullName:
          type: string

    Library:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string

    BookId:
      type: integer
      format: int64