# library-backend/src/main/java/com/lbdg/library_backend/DTOs/responseDTOs/BookingEditResponseDTO.java
package com.lbdg.library_backend.DTOs.responseDTOs;

import com.fasterxml.jackson.annotation.JsonProperty;
import lombok.Builder;
import lombok.Value;
import lombok.extern.jackson.Jacksonized;

import java.time.LocalDate;

@Builder
@Jacksonized
@Value
public class BookingEditResponseDTO {
    @JsonProperty("userId")
    private Long userId;
    @JsonProperty("startDate")
    private LocalDate startDate;
    @JsonProperty("endDate")
    private LocalDate endDate;
}